0.4.1
Added Telegram notifications that are sent when the ESP is started, the current mode of the inverter is changed, the battery is low (<15%), the battery is fully charged (100%), the heatsink temperature is above 35 degrees, and cSOC changed.

In the settings section, fields have been added for the Telegram bot token and chat ID to send notifications to.

Сhanged path for bin firmware files